+++ src/recados/recados.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Recado } from './entities/recado.entity';
 import { CreateRecadoDto } from './dto/create-recado.dto';
 import { UpdateRecadoDto } from './dto/update-recado.dto';
@@ -7,6 +11,7 @@ import { Repository } from 'typeorm';
 import { PessoasService } from 'src/pessoas/pessoas.service';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
 import { RecadosUtils } from './recados.utils';
+import { TokenPayloadDto } from 'src/auth/dto/token-payload.dto';
 
 @Injectable()
 export class RecadosService {
@@ -73,11 +78,15 @@ export class RecadosService {
     if (recado) return recado;
   }
 
-  async create(createRecadoDto: CreateRecadoDto) {
-    const { deId, paraId } = createRecadoDto;
+  async create(
+    createRecadoDto: CreateRecadoDto,
+    tokenPayload: TokenPayloadDto,
+  ) {
+    const { paraId } = createRecadoDto;
 
-    const de = await this.pessoasService.findOne(deId);
     const para = await this.pessoasService.findOne(paraId);
+
+    const de = await this.pessoasService.findOne(tokenPayload.sub);
 
     const newRecado = {
       texto: createRecadoDto.texto,
@@ -91,13 +100,21 @@ export class RecadosService {
     await this.recadoRepository.save(recado);
     return {
       ...recado,
-      de: { id: recado.de.id },
-      para: { id: recado.para.id },
+      de: { id: recado.de.id, nome: recado.de.nome },
+      para: { id: recado.para.id, nome: recado.para.nome },
     };
   }
 
-  async update(id: number, updateRecadoDto: UpdateRecadoDto) {
+  async update(
+    id: number,
+    updateRecadoDto: UpdateRecadoDto,
+    tokenPayload: TokenPayloadDto,
+  ) {
     const recado = await this.findOne(id);
+
+    if (recado?.de.id !== tokenPayload.sub) {
+      throw new ForbiddenException('Esse recado não é seu!');
+    }
 
     if (!recado) {
       this.throwNotFoundError();
@@ -109,12 +126,16 @@ export class RecadosService {
     }
   }
 
-  async remove(id: number) {
-    const recado = await this.recadoRepository.findOneBy({ id });
+  async remove(id: number, tokenPayload: TokenPayloadDto) {
+    const recado = await this.findOne(id);
 
     if (!recado) {
       this.throwNotFoundError();
       return;
+    }
+
+    if (recado?.de.id !== tokenPayload.sub) {
+      throw new ForbiddenException('Esse recado não é seu!');
     }
 
     return this.recadoRepository.remove(recado);
